[#73] feat: 인원 정보 수정 함수 추가
- id로 인원 정보 조회 함수 위치를 변경합니다.
